+++ Cesium3DTilesSelection/include/Cesium3DTilesSelection/Tileset.h
@@ -522,10 +522,12 @@ private:
       const TilesetFrameState& frameState,
       Tile& tile,
       double tilePriority,
+      double tileSse,
       ViewUpdateResult& result);
   TraversalDetails _renderInnerTile(
       const TilesetFrameState& frameState,
       Tile& tile,
+      double tileSse,
       ViewUpdateResult& result);
   bool _kickDescendantsAndRenderTile(
       const TilesetFrameState& frameState,
@@ -535,7 +537,8 @@ private:
       size_t firstRenderedDescendantIndex,
       const TilesetViewGroup::LoadQueueCheckpoint& loadQueueBeforeChildren,
       bool queuedForLoad,
-      double tilePriority);
+      double tilePriority,
+      double tileSse);
   TileOcclusionState _checkOcclusion(const Tile& tile);
 
   TraversalDetails _visitTile(
@@ -545,6 +548,7 @@ private:
       bool ancestorMeetsSse,
       Tile& tile,
       double tilePriority,
+      double tileSse,
       ViewUpdateResult& result);
 
   struct CullResult {
@@ -564,11 +568,11 @@ private:
       const TilesetFrameState& frameState,
       const std::vector<double>& distances,
       CullResult& cullResult);
-  bool _meetsSse(
+  double _computeSse(
       const std::vector<ViewState>& frustums,
       const Tile& tile,
-      const std::vector<double>& distances,
-      bool culled) const noexcept;
+      const std::vector<double>& distances) const noexcept;
+  bool _meetsSseThreshold(double sse, bool culled) const noexcept;
 
   TraversalDetails _visitTileIfNeeded(
       const TilesetFrameState& frameState,
@@ -593,6 +597,7 @@ private:
    * @param result The current view update result.
    * @param tilePriority The load priority of this tile.
    * priority.
+   * @param tileSse The screen space error of this tile.
    * @param queuedForLoad True if this tile has already been queued for loading.
    * @return true The additive-refined tile was queued for load and added to the
    * render list.
@@ -603,6 +608,7 @@ private:
       Tile& tile,
       ViewUpdateResult& result,
       double tilePriority,
+      double tileSse,
       bool queuedForLoad);
 
   void _unloadCachedTiles(double timeBudget) noexcept;